refactor: 파티를 스텝 3로 분리
